+++ datumaro/components/dataset.py
@@ -19,13 +19,15 @@ from datumaro.components.dataset_filter import (
 )
 from datumaro.components.environment import Environment
 from datumaro.components.errors import (
-    CategoriesRedefinedError, MultipleFormatsMatchError, NoMatchingFormatsError,
-    RepeatedItemError, UnknownFormatError,
+    CategoriesRedefinedError, ConflictingCategoriesError,
+    MultipleFormatsMatchError, NoMatchingFormatsError, RepeatedItemError,
+    UnknownFormatError,
 )
 from datumaro.components.extractor import (
     DEFAULT_SUBSET_NAME, CategoriesInfo, DatasetItem, Extractor, IExtractor,
     ItemTransform, Transform,
 )
+from datumaro.plugins.transforms import ProjectLabels
 from datumaro.util import is_method_redefined
 from datumaro.util.log_utils import logging_disabled
 from datumaro.util.os_util import rmtree
@@ -552,7 +554,7 @@ class DatasetStorage(IDataset):
         self._length = None
 
     def has_updated_items(self):
-        return self._transforms or self._updated_items
+        return bool(self._transforms) or bool(self._updated_items)
 
     def get_patch(self):
         # Patch includes only added or modified items.
@@ -572,14 +574,33 @@ class DatasetStorage(IDataset):
             else:
                 patch.put(self._storage.get(item_id, subset))
 
-        return DatasetPatch(patch, self._categories,
-            self._updated_items)
+        return DatasetPatch(patch, self._categories, self._updated_items)
 
     def flush_changes(self):
         self._updated_items = {}
         if not (self.is_cache_initialized() or self._is_unchanged_wrapper):
             self._flush_changes = True
 
+    def update(self,
+            source: Union[DatasetPatch, IExtractor, Iterable[DatasetItem]]):
+        # TODO: provide a more efficient implementation with patch reuse
+
+        if isinstance(source, DatasetPatch):
+            if source.categories() != self.categories():
+                raise ConflictingCategoriesError()
+
+            for item_id, status in source.updated_items.items():
+                if status == ItemStatus.removed:
+                    self.remove(*item_id)
+                else:
+                    self.put(source.data.get(*item_id))
+        elif isinstance(source, IExtractor):
+            for item in ProjectLabels(source, self.categories().get(
+                    AnnotationType.label, LabelCategories())):
+                self.put(item)
+        else:
+            for item in source:
+                self.put(item)
 
 class Dataset(IDataset):
     _global_eager = False
@@ -691,9 +712,22 @@ class Dataset(IDataset):
         else:
             return self.transform(XPathDatasetFilter, expr)
 
-    def update(self, items: Iterable[DatasetItem]) -> 'Dataset':
-        for item in items:
-            self.put(item)
+    def update(self,
+            source: Union[DatasetPatch, IExtractor, Iterable[DatasetItem]]) \
+                -> 'Dataset':
+        """
+        Updates items of the current dataset from another dataset or an
+        iterable (the source). Items from the source overwrite matching
+        items in the current dataset. Unmatched items are just appended.
+
+        If the source is a DatasetPatch, the removed items in the patch
+        will be removed in the current dataset.
+
+        If the source is a dataset, labels are matched. If the labels match,
+        but the order is different, the annotation labels will be remapped to
+        the current dataset label order during updating.
+        """
+        self._data.update(source)
         return self
 
     def transform(self, method: Union[str, Type[Transform]],
@@ -793,6 +827,9 @@ class Dataset(IDataset):
 
     @scoped
     def export(self, save_dir: str, format, **kwargs):
+        if not save_dir:
+            raise ValueError("Dataset export path is not specified")
+
         inplace = (save_dir == self._source_path and format == self._format)
 
         if isinstance(format, str):