+++ datumaro/components/project.py
@@ -31,11 +31,12 @@ from datumaro.components.errors import (
     DatasetMergeError, EmptyCommitError, EmptyPipelineError,
     ForeignChangesError, InvalidStageError, MigrationError,
     MismatchingObjectError, MissingObjectError, MissingPipelineHeadError,
-    MultiplePipelineHeadsError, OldProjectError, PathOutsideSourceError,
-    ProjectAlreadyExists, ProjectNotFoundError, ReadonlyDatasetError,
-    ReadonlyProjectError, SourceExistsError, SourceUrlInsideProjectError,
-    UnexpectedUrlError, UnknownRefError, UnknownSourceError, UnknownStageError,
-    UnknownTargetError, UnsavedChangesError, VcsError,
+    MissingSourceHashError, MultiplePipelineHeadsError, OldProjectError,
+    PathOutsideSourceError, ProjectAlreadyExists, ProjectNotFoundError,
+    ReadonlyDatasetError, ReadonlyProjectError, SourceExistsError,
+    SourceUrlInsideProjectError, UnexpectedUrlError, UnknownRefError,
+    UnknownSourceError, UnknownStageError, UnknownTargetError,
+    UnsavedChangesError, VcsError,
 )
 from datumaro.components.launcher import Launcher
 from datumaro.util import find, parse_str_enum_value
@@ -62,7 +63,8 @@ class ProjectSourceDataset(IDataset):
         self.__dict__['name'] = source
 
     def save(self, save_dir=None, **kwargs):
-        if save_dir is None and self.readonly:
+        if self.readonly and (save_dir is None or \
+                osp.abspath(save_dir) == osp.abspath(self.data_path)):
             raise ReadonlyDatasetError()
         self._dataset.save(save_dir, **kwargs)
 
@@ -348,13 +350,16 @@ class ProjectBuilder:
                     "is disabled in read-only projects.", source_name)
                 continue
 
-            # TODO: check if we can avoid computing source hash in some cases
-            assert source.hash, source_name
+            if not source.hash:
+                raise MissingSourceHashError("Unable to re-download source "
+                    "'%s': the source was added with no hash information. " % \
+                    source_name)
+
             with self._project._make_tmp_dir() as tmp_dir:
                 obj_hash, _, _ = \
                     self._project._download_source(source.url, tmp_dir)
 
-                if source.hash != obj_hash:
+                if source.hash and source.hash != obj_hash:
                     raise MismatchingObjectError(
                         "Downloaded source '%s' data is different " \
                         "from what is saved in the build pipeline: "