+++ tests/test_project.py
@@ -11,9 +11,9 @@ from datumaro.components.config_model import Model, Source
 from datumaro.components.dataset import DEFAULT_FORMAT, Dataset
 from datumaro.components.errors import (
     DatasetMergeError, EmptyCommitError, ForeignChangesError,
-    MismatchingObjectError, MissingObjectError, OldProjectError,
-    PathOutsideSourceError, ReadonlyProjectError, SourceExistsError,
-    SourceUrlInsideProjectError,
+    MismatchingObjectError, MissingObjectError, MissingSourceHashError,
+    OldProjectError, PathOutsideSourceError, ReadonlyProjectError,
+    SourceExistsError, SourceUrlInsideProjectError,
 )
 from datumaro.components.extractor import DatasetItem, Extractor, ItemTransform
 from datumaro.components.launcher import Launcher
@@ -1011,6 +1011,27 @@ class ProjectTest(TestCase):
         self.assertNotEqual('', project.working_tree.sources['source1'].hash)
         self.assertNotEqual('',
             project.working_tree.build_targets['source1'].head.hash)
+
+    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
+    @scoped
+    def test_cant_redownload_unhashed(self):
+        test_dir = scope_add(TestDir())
+        dataset_url = osp.join(test_dir, 'dataset')
+        Dataset.from_iterable([
+            DatasetItem('a'),
+            DatasetItem('b'),
+        ]).save(dataset_url)
+
+        proj_dir = osp.join(test_dir, 'proj')
+        project = scope_add(Project.init(proj_dir))
+        project.import_source('source1', url=dataset_url,
+            format=DEFAULT_FORMAT, no_hash=True)
+        project.working_tree.build_targets.add_transform_stage('source1',
+            'reindex')
+        project.commit('a commit')
+
+        with self.assertRaises(MissingSourceHashError):
+            project.working_tree.make_dataset('source1.root')
 
 class BackwardCompatibilityTests_v0_1(TestCase):
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)